Serious refactor of the object concept
Now an object is made up of many different areas that can be of
different sizes, and the our detection algorithm now works with any
arbitrary shape.

Currently, only the Wall object works, but the ArenaField now has a
single bordered wall that renders well with a single stroke instead of
the obvious, multiple rectangles.